Move vendored MurmurHash3 into ext/.
MurmurHash3 is third-party (public domain). Placing it under
src/search/ext/ excludes it from the clang-format and include-guard
style checks, which flagged its upstream formatting and _MURMURHASH3_H_
guard. Update the include in cost_saturation/utils.h and the CMake
source path.
