+++ src/search/cost_saturation/utils.h
@@ -4,7 +4,7 @@
 #include "abstraction.h"
 #include "types.h"
 
-#include "../utils/murmurhash3.h"
+#include "../ext/murmurhash3/murmurhash3.h"
 
 #include <cstdint>
 #include <execution>
